requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
waitress==3.0.2
psycopg2-binary==2.9.7